Refactor svmbir.py and put interface code into py_c_interface.py
